+++ dolomite_engine/hf_models/modeling_utils_TP/TP.py
@@ -4,7 +4,7 @@ import torch
 import torch.distributed
 import torch.nn as nn
 from torch.distributed._tensor.api import DTensor
-from torch.distributed._tensor.placement_types import Placement, Replicate, Shard
+from torch.distributed._tensor.placement_types import Placement
 from torch.profiler import record_function
 
 from ...utils import ProcessGroupManager
@@ -136,37 +136,29 @@ def tensor_parallel_split_safetensor_slice(slice, dim: int, start_end: Tuple[int
         raise RuntimeError("this code should not be reachable")
 
 
-def tensor_to_dtensor_hook(
-    module: nn.Module, inputs: tuple[torch.Tensor], current_placement: Placement, desired_placement: Placement = None
+def tensor_to_dtensor(
+    tensor: torch.Tensor, current_placement: Placement, desired_placement: Placement = None
 ) -> DTensor:
-    assert len(inputs) == 1
-    input = inputs[0]
-
     tp_mesh = ProcessGroupManager.get_tensor_parallel_mesh()
 
-    input = DTensor.from_local(input, device_mesh=tp_mesh, run_check=False, placements=[current_placement])
-
+    dtensor = DTensor.from_local(tensor, device_mesh=tp_mesh, run_check=False, placements=[current_placement])
     if desired_placement is not None:
-        input = input.redistribute(device_mesh=tp_mesh, placements=[desired_placement])
+        dtensor = dtensor.redistribute(device_mesh=tp_mesh, placements=[desired_placement])
 
-    return (input,)
+    return dtensor
 
 
-def dtensor_to_tensor_hook(
-    module: nn.Module,
-    inputs: tuple[DTensor],
-    output: DTensor,
-    desired_placement: Placement = None,
-    grad_placement: Placement = None,
+def dtensor_to_tensor(
+    dtensor: DTensor, desired_placement: Placement = None, grad_placement: Placement = None
 ) -> torch.Tensor:
     if desired_placement is not None:
-        output = output.redistribute(
+        dtensor = dtensor.redistribute(
             device_mesh=ProcessGroupManager.get_tensor_parallel_mesh(), placements=[desired_placement]
         )
 
-    output = output.to_local(grad_placements=None if grad_placement is None else [grad_placement])
+    tensor = dtensor.to_local(grad_placements=None if grad_placement is None else [grad_placement])
 
-    return output
+    return tensor
 
 
 @torch.no_grad()
