+++ dolomite_engine/hf_models/modeling_utils_TP/embedding.py
@@ -11,7 +11,8 @@ from torch.distributed._tensor.placement_types import Replicate, Shard
 from ...utils import ProcessGroupManager, SafeTensorsWeightsManager, get_cuda_rng_tracker
 from ..modeling_utils import ParameterizedEmbedding
 from ..utils import divide_if_divisible
-from .TP import dtensor_to_tensor_hook, modify_state_dict_to_dtensor_dict, tensor_to_dtensor_hook
+from .hooks import dtensor_to_tensor_hook, tensor_to_dtensor_hook
+from .TP import modify_state_dict_to_dtensor_dict
 
 
 class Embedding_TP(ParameterizedEmbedding):
